pyproject.toml
[tool.poetry]
name = "bloganalysis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.12.2"
matplotlib = "^3.8.2"
pytest = "^7.4.4"
openai = "^1.7.2"
tenacity = "^8.2.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
